refactor default.nix
This breaks backward-compatibility completely.

Renames for consistency:
* generateYarnNix -> mkYarnNix
* loadOfflineCache -> importOfflineCache
* buildYarnPackage -> mkYarnPackage
* buildYarnPackageDeps -> mkYarnModules

mkYarnPackage has been overhauled to be more flexible and support
various scenarios. Default options are also provided for more attributes
to minimize the boilerplate for package writers.

Here is a frontend package example:

    mkYarnPackage {
      src = ./.;
      # package.json and yarn.lock paths are inferred from src
      # name is inferred from package.json

      buildPhase = ''
        # generates the dist/ folder from the source
        yarn build
      '';

      # this replaces the default phase that links the .bin folder
      installPhase = ''
        cp -r dist/ $out
      '';
    }

As far as I know this supports all the use-case scenarios that I am
aware of.