Adds support for org names in package.json
This commit adds a function to reformat the string in the name field of
a package.json into a format that nix can use as a derivation name. This
is done through a combination of nix builtins that may not be ideal.

This is necessary because the organization in npm scoping is incompatible with
the format for nix package names.