remove mkYarnPackage
The Nix file for mkYarnPackage must be in the nixpkgs repository,
because Hydra does not allow importing Nix files from downloaded paths.
To prevent the nixpkgs version of the Nix code from getting out-of-sync
with this repository, let's remove it here and use nixpkgs as the single
source of truth for the Nix code.
Incompatibilities in the interface can be dealt with by adding an
attribute for the yarn.nix file format version in the future.